requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
pillow
PyPDF2
python-docx
jinja2
gunicorn
stripe
python-dotenv
bcrypt
flask-wtf
wtforms
email-validator